Issue 0021247: EDF 1856 STUDY: Unification of the configuration files and directories of Salome Part 1: Store GUI prefrences (+ style) in the ~/.config/salome directory

+++ src/SUIT/SUIT_ResourceMgr.h
@@ -41,7 +41,7 @@ public:
 protected:
   virtual QString userFileName( const QString&, const bool = true ) const;
   virtual QString findAppropriateUserFile( const QString& ) const;
-  virtual int     userFileId( const QString& ) const;
+  virtual long    userFileId( const QString& ) const;
 
 private:
   QString         myVersion;
